+++ src/main/java/com/singlestore/jdbc/plugin/AuthenticationPlugin.java
@@ -2,10 +2,8 @@
 // Copyright (c) 2012-2014 Monty Program Ab
 // Copyright (c) 2015-2024 MariaDB Corporation Ab
 // Copyright (c) 2021-2024 SingleStore, Inc.
-
 package com.singlestore.jdbc.plugin;
 
-import com.singlestore.jdbc.Configuration;
 import com.singlestore.jdbc.client.Context;
 import com.singlestore.jdbc.client.ReadableByteBuf;
 import com.singlestore.jdbc.client.socket.Reader;
@@ -14,22 +12,6 @@ import java.io.IOException;
 import java.sql.SQLException;
 
 public interface AuthenticationPlugin {
-
-  /**
-   * Authentication plugin type.
-   *
-   * @return authentication plugin type. ex: mysql_native_password
-   */
-  String type();
-
-  /**
-   * Plugin initialization.
-   *
-   * @param authenticationData authentication data (password/token)
-   * @param seed server provided seed
-   * @param conf Connection options
-   */
-  void initialize(String authenticationData, byte[] seed, Configuration conf);
 
   /**
    * Process plugin authentication.